Unify parsers and refactor to AbstractParser (task #3495)
* All parsers now work with JSON schema
* All parsers now support defaults
* Parsing is generic between parsers, but fetching file content
  is not.
