+++ src/ModuleConfig/Parser/Csv/AbstractCsvParser.php
@@ -1,8 +1,10 @@
 <?php
 namespace Qobo\Utils\ModuleConfig\Parser\Csv;
 
+use Exception;
 use League\Csv\Reader;
 use Qobo\Utils\ModuleConfig\Parser\AbstractParser;
+use Qobo\Utils\Utility;
 
 abstract class AbstractCsvParser extends AbstractParser
 {
@@ -21,10 +23,37 @@ abstract class AbstractCsvParser extends AbstractParser
     {
         $result = [];
 
-        $this->validatePath($path);
+        try {
+            Utility::validatePath($path);
+        } catch (Exception $e) {
+            return $result;
+        }
 
         $reader = Reader::createFromPath($path, $this->open_mode);
         $result = $reader->fetchOne();
+
+        return $result;
+    }
+
+    /**
+     * Read and parse a given path
+     *
+     * @return array
+     */
+    protected function getDataFromPath($path)
+    {
+        $result = [];
+
+        // If no structure specified (default or param), then use headers
+        if (empty($this->options['structure'])) {
+            $this->options['structure'] = $this->getHeadersFromPath($path);
+        }
+
+        $reader = Reader::createFromPath($path, $this->open_mode);
+        $rows = $reader->setOffset(1)->fetchAssoc($this->options['structure']);
+        foreach ($rows as $row) {
+            $result[] = $row;
+        }
 
         return $result;
     }
