+++ src/ModuleConfig/Parser/Csv/ListParser.php
@@ -9,8 +9,19 @@ namespace Qobo\Utils\ModuleConfig\Parser\Csv;
  *
  * @author Leonid Mamchenkov <l.mamchenkov@qobo.biz>
  */
-class ListParser extends Parser
+class ListParser extends AbstractCsvParser
 {
+    /**
+     * JSON schema
+     *
+     * This can either be a string, pointing to the file
+     * or an StdClass with an instance of an already parsed
+     * schema
+     *
+     * @var string|StdClass $schema JSON schema
+     */
+    protected $schema = 'file://' . __DIR__ . DIRECTORY_SEPARATOR . '..' . DIRECTORY_SEPARATOR . 'Schema' . DIRECTORY_SEPARATOR . 'list.json';
+
     /**
      * Parsing options
      */
