+++ src/ModuleConfig/Parser/Csv/MigrationParser.php
@@ -9,8 +9,19 @@ namespace Qobo\Utils\ModuleConfig\Parser\Csv;
  *
  * @author Leonid Mamchenkov <l.mamchenkov@qobo.biz>
  */
-class MigrationParser extends Parser
+class MigrationParser extends AbstractCsvParser
 {
+    /**
+     * JSON schema
+     *
+     * This can either be a string, pointing to the file
+     * or an StdClass with an instance of an already parsed
+     * schema
+     *
+     * @var string|StdClass $schema JSON schema
+     */
+    protected $schema = 'file://' . __DIR__ . DIRECTORY_SEPARATOR . '..' . DIRECTORY_SEPARATOR . 'Schema' . DIRECTORY_SEPARATOR . 'migration.json';
+
     /**
      * Parsing options
      */
@@ -44,11 +55,6 @@ class MigrationParser extends Parser
         $fields = parent::parse($path, $options);
         if (empty($fields)) {
             return $fields;
-        }
-
-        // Overwrite defaults
-        if (!empty($options)) {
-            $this->options = $options;
         }
 
         // Convert indexed array to associative one,
